refactor: move insight savings rate types
