+++ frontend/src/insights/types/savingsRate.ts
@@ -1,0 +1,10 @@
+export type SavingsRateHistoryPoint = {
+  monthKey: string
+  monthLabel: string
+  tickLabel: string
+  fullLabel: string
+  rate: number | null
+  income: number
+  expenses: number
+  isCurrent: boolean
+}
